Return response of beforeDestroy if any

+++ src/Controllers/Traits/Crud/Destroy.php
@@ -91,7 +91,9 @@ trait Destroy
 
         return DB::transaction(
             function () use (&$item) {
-                $this->beforeDestroy($item);
+                if ($resp = $this->beforeDestroy($item)) {
+                    return $resp;
+                }
                 $result = $item->delete();
     
                 if (!$result) {
@@ -170,7 +172,9 @@ trait Destroy
         }
         return DB::transaction(
             function () use (&$data, $model) {
-                $this->beforeDestroyMany($data);
+                if ($resp = $this->beforeDestroyMany($data)) {
+                    return $resp;
+                }
         
                 $result = is_object($model)
                     ? $model->whereIn('id', $data['ids'])->delete()
@@ -255,7 +259,9 @@ trait Destroy
 
         return DB::transaction(
             function () use (&$item) {
-                $this->beforeForceDestroy($item);
+                if ($resp = $this->beforeForceDestroy($item)) {
+                    return $resp;
+                }
                 $result = $item->forceDelete();
         
                 if (!$result) {
@@ -338,7 +344,9 @@ trait Destroy
 
         return DB::transaction(
             function () use (&$item) {
-                $this->beforeRestoreDestroyed($item);
+                if ($resp = $this->beforeRestoreDestroyed($item)) {
+                    return $resp;
+                }
                 $result = $item->restore();
         
                 if (!$result) {
